Rename plugin.CatalogImpl as plugin.CatalogPlugin
